Split panes UI components
